// src/components/ArcGIS/types/stream.ts
export type Stream = {
    major_stream_legacy_code: string|null
    major_stream_name: string|null
    legacy_code: string|null
    legacy_name: string|null
    OBJECTID: string|null
    stream_id: number|null
    stream_name: string|null
    mhfd_code_stream: string|null
    max_catchment_sum_ac: number|null
    sum_stream_length_ft: number|null
    sum_stream_length_miles: number|null
    sum_catchment_area_ac: number|null
    reach_code: string|null
    trib_code1: number|null
    trib_code2: number|null
    trib_code3: number|null
    trib_code4: number|null
    trib_code5: number|null
    trib_code6: number|null
    trib_code7: number|null
    created_user: string|null
    created_date: string|null
    last_edited_user: string|null
    last_edited_date: string|null
    Shape__Length: number|null
}